Step 38 実装: エンティティ・武器の ID ベース参照

+++ native/game_native/benches/ai_bench.rs
@@ -14,7 +14,7 @@ fn setup_enemies(n: usize) -> EnemyWorld {
             (x, y)
         })
         .collect();
-    enemies.spawn(&positions, EnemyKind::Slime);
+    enemies.spawn(&positions, EnemyKind::Slime as u8);
     enemies
 }
